requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-filter==22.1
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==1.3
ecdsa==0.17.0
pyasn1==0.4.8
PyJWT==2.4.0
python-jose==3.3.0
pytz==2022.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
